Refactor CasesIter into separate file
